Reduce move speed and add update_pose TODO: complete pose update using odometry

+++ Hardware/include-scope/hardware-scope.h
@@ -28,7 +28,7 @@ enum HardwareConstants {
 struct Pose {
   int x;
   int y;
-  int theta;
+  float theta;
 };
 
 }  // namespace hardware
